SR-034 Finished interests filter UI [SD,PO]

+++ app/src/main/java/at/team30/setroute/models/FilteringOptions.kt
@@ -2,12 +2,12 @@ package at.team30.setroute.models
 
 class FilteringOptions (
     var interests: List<Int> = listOf(0, 1, 2, 3, 4, 5, 6),
-    var minDistance: Int = 0,
-    var maxDistance: Int = 20,
-    var minDuration: Int = 0,
-    var maxDuration: Int = 120) {
+    var minDistance: Float = 0f,
+    var maxDistance: Float = 20f,
+    var minDuration: Float = 0f,
+    var maxDuration: Float = 120f) {
     companion object {
-        fun fromValues(interests: List<Int>, minDistance: Int, maxDistance: Int, minDuration: Int, maxDuration: Int): FilteringOptions {
+        fun fromValues(interests: List<Int>, minDistance: Float, maxDistance: Float, minDuration: Float, maxDuration: Float): FilteringOptions {
             return FilteringOptions(interests, minDistance, maxDistance, minDuration, maxDuration)
         }
     }
