+++ app/src/main/java/at/team30/setroute/ui/routes/RouteListViewModel.kt
@@ -66,7 +66,7 @@ class RouteListViewModel @Inject constructor(
 
     }
 
-    fun applyFiltering(interests: List<Int>, minDistance: Int, maxDistance: Int, minDuration: Int, maxDuration: Int) {
+    fun applyFiltering(interests: List<Int>, minDistance: Float, maxDistance: Float, minDuration: Float, maxDuration: Float) {
         filteringRepository.storeFilteringOptions(FilteringOptions.fromValues(interests, minDistance, maxDistance, minDuration, maxDuration))
 
         //routesList.postValue(getFilteredRoutes())
